chore(ECO-3155): Move arena utils to a more sensible location
