+++ src/typescript/sdk/src/markets/index.ts
@@ -1,2 +1,1 @@
-export * from "./arena-utils";
 export * from "./utils";
